+++ react/components/EditorContainer/Sidebar/BlockEditor/BlockConfigurationEditor/utils.ts
@@ -8,7 +8,6 @@ import {
   getExtension,
   getIframeImplementation,
 } from '../../../../../utils/components'
-
 import { GetSchemasArgs } from './typings'
 
 /**