+++ react/components/EditorContainer/Sidebar/BlockSelector/BlockList/BlockListItem/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 
 import { NormalizedBlock } from '../../typings'
-
 import ExpandArrow from './ExpandArrow'
 import Item from './Item'