+++ react/components/EditorContainer/Sidebar/index.tsx
@@ -72,7 +72,7 @@ const Sidebar: React.FunctionComponent<Props> = ({
   return (
     <div
       id="sidebar-vtex-editor"
-      className="z-1 h-100 w-100 flex flex-row-reverse"
+      className={'z-1 h-100 w-100 flex flex-row-reverse ' + scrollState}
     >
       <nav
         ref={adminSidebar}
@@ -83,7 +83,7 @@ const Sidebar: React.FunctionComponent<Props> = ({
           scrollState
         }
       >
-        <div className="relative h-100 flex flex-column dark-gray">
+        <div className="relative h-100 flex flex-column dark-gray overflow-y-auto">
           <Modal
             isActionDanger={isActionDanger}
             isActionLoading={editor.getIsLoading()}