+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Status toggle for content configurations.
+
 ## [4.44.2] - 2021-07-12
 
 ### Fixed